nvim: disable ghost_text in cmdline completion
The '/' key is bound to accept the current selection to make tabing out
directories and sub directories more natural. When ghost-text is enabled
in the cmdline, this interferes with commands like `:s`, where it is
also natural to type ":%s/". With ghost-text and this key bind, this
leads to incorrect autocompletions, e.g. ':%s/' -> ':%sall'.